refactor: 트립페이지 모바일 버전 컴포넌트 분리
